+++ CHANGELOG.md
@@ -16,6 +16,18 @@ especially the [Rust flavour](https://doc.rust-lang.org/cargo/reference/semver.h
 
 ### Removed
 
+## [0.1.1] - 2025-??-??
+
+### Added
+- possibility to override the first tick: async fn start(...) -> BehaviorResult
+- possibility to override the halt method: fn halt(...) -> Result<(), BehaviorError>
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [0.1.0] - 2025-07-29
 
 Version 0.1.0 is an implementation of the core features of [BehaviorTree.CPP](https://www.behaviortree.dev/).